// src/types/interfaces.ts
export interface IAction {
  type: string;
  payload: any;
}

export interface IState {
  showNav: boolean;
  dark: boolean;
  shows: Array<IShow>;
  episodes: Array<IEpisode>;
  favourites: Array<IEpisode>;
  visualisations: Array<IVis>;
  yahooFinanceApiOff: boolean;
  yahooFinanceApiOffChartWithinCard: boolean;
  yahooFinanceApiOffStockCard: boolean;
  sensorWebocketsOff: boolean;
  auth: IAuth | null;
  user: IUser | null;
}

export interface IAuth {
  accessToken: string;
  idToken: string;
  isAuthenticated: boolean;
  isPending: boolean;
}

export interface IUser {
  email: string;
  email_verified: boolean;
  family_name: string;
  given_name: string;
  locale: string;
  name: string;
  preferred_username: string;
  sub: string;
  updated_at: number;
  zoneinfo: string;
}

export interface IShow {
  score: number;
  show: {
    id: number;
    url: string; //www.tvmaze.com/shows/216/rick-and-morty;
    name: string;
    type: string;
    language: string;
    genres: Array<string>;
    status: string;
    runtime: number;
    premiered: string;
    officialSite: string;
    schedule: {
      time: string;
      days: Array<string>;
    };
    rating: {
      average: number;
    };
    weight: number;
    network: {
      id: number;
      name: string;
      country: {
        name: string;
        code: string;
        timezone: string;
      };
    };
    webChannel: string;
    externals: {
      tvrage: number;
      thetvdb: number;
      imdb: string;
    };
    image: {
      medium: string;
      original: string;
    };
    summary: HTMLElement;
    updated: number;
    _links: {
      self: {
        href: string;
      };
      previousepisode: {
        href: string;
      };
    };
  };
}

export interface IEpisode {
  airdate: string;
  airstamp: string;
  airtime: string;
  id: number;
  image: {
    medium: string;
    original: string;
  };
  name: string;
  number: number;
  runtime: number;
  season: number;
  summary: string;
  url: string;
}

export interface IEpisodeProps {
  episodes: Array<IEpisode>;
  toggleFavAction: (state: IState, dispatch: any, episode: IEpisode) => IAction;
  store: { state: IState; dispatch: any };
  favourites: Array<IEpisode>;
}

export type Dispatch = React.Dispatch<IAction>;

export interface IVis {
  createdDate: string;
  id: number;
  mediaType: "video" | "img";
  media: {
    medium: string;
    original: string;
  };
  name: string;
  summary: string;
  link: string;
}